[WIP] Create php interface to mysql server
This is still a work in progress. This tool should allow us to
communicate with the mysql server very easily with minimal thought of
sql syntax. The intent is that this tool will allow us to dynamically
generate sql commands and queries to send to the server. This will be
very useful when receiving fields to filter on from the webpage.

Change-Id: I9755aad6234e0adda37234af7651d906b7e1229c