Check for presence of the database before filter submit
So this is my first effor to try and make a proper commit message.
A file, php/has_db.php gets added in all it's glory, and it simply
has a function that takes a name of a database and checks if it is
present in the local mysql system of the server. If it is not,
js/main.html then resubmits the PGN file before sending the filters
to perform the approcpriate queries.

This process must happen synchronously to be able to work properly,
otherwise race conditions will emerge, and considering South African
internet speeds, they WILL occur.

Change-Id: Ifdaebe586955fc0de61df7f3d824c071c6caa67d