Change pgn_parser to open & close db outside while
Currently, the parser opens and closes access to the database and
selects the current database for every single chess game being added.
For a 2.5 million game pgn file, that ammounts to about 7.5 million
additional queries, that could be reduced to three.

So, this should be a nice optimization, unless there is some specific
reason to open and close the db everytime, like flushing of memory,
for instance? Let me know what you think.

Change-Id: I0db110a94007c418bce554c9a19307a6425d5e0e