Strip js/sha256.js of all unused statements
The sha256 library we include is an open source library that has
various different implementations, depending on the exact output the
user might want. We are using exactly one of the options at this
point, making more than half the library unnecessary.  Stripping these
out makes a big difference in the load time on slow connections, at no
functional disadvantage to the project.

Furthermore, I removed the unused (unminified) bootstrap library, since
bootstrap.min.js is the version we use and we are unlikely to directly
want to alter this file ever.

Change-Id: I789cf920066102e8e4b8a5c7e0c0e4fd285407a0