Display a d3 graph
Data received from php/query.php is stored in a global variable in
js/main.js, once this data has been processed by js/visual_response.js
 it stored in another global variable and the original data is
hashed an stored into another global variable. This is done to
optimize the responsiveness of the graph, which is redrawn on the
resizing of the window.

Change-Id: I962a0756b525c3f3faa7c5275d3827fc1402140d