Add a developer mode for easy local testing
This is actually a bit of a selfish change as I think it will mainly
help me. But it that said, it would help me quite a lot. It basically
adds a developer moad field to the config file and if that is set to
true the javascripts sends all requests to the local machine.

Change-Id: If34031bb61110b9dce14a154cc55c8d53c5bc931