Give choice of query -- eco tag vs eco category
Currently we just filter on eco tag, of which there are literally a
a total of 500.  This data is to finely grained.  Therefore, I think it is
more appropriate to filter on the 13 categories by default.

With this change, this is now the case, and users can opt for the finely
graned option by setting a number in one of the eco range filter spaces.

Change-Id: I53496a6f7eadb88424776955cae479629e426e81