Update the parser to exclude eco tag-less games
Currently, if a game in the pgn file doesn't have an ECO tag, it's eco
class is omitted and the numeric part is stored as -1.  There are 1544 such
games, which is 0.1% of the total, all between 2000 and 2015.  As we
literally filter and group on these values, in most cases these data will
not be included in the filtered results.  It is also a small enough number
that it doesn't make a huge statistical difference.

I emailed Brink van der Merwe and he agrees it is reasonable to just
exclude these games from the database.  As such, I have done just that in
the commit.

Change-Id: I018892818543482bd448441c297051c22cc9253d