Make a new query file for the final kind of query
This file is still very crude, with many comments to be taken out and such,
but I want your feedback so long.

php/elo_query.php is basically identical in structure to php/query.php, but
it has fewer filters that the user can control, has a whole different query
and the results must be processed more intensely.

Change-Id: I74925d0bcf7ef98ff9576074de2b77fd4ad64a63