Refine query options
So, first of all, there was a discordance between what the database
considers a "category" and what the front end considered a "category".
Hence, I have updated the front end to be more consistent.  I believed this
reasonable, as the specifications provided are unclear on this anyway.

Before I continue, as of this commit, an eco code here refers to the tag
provided in pgn, ie, "A11"; an eco class refers to the alphabetic portion,
so "A11" would be in class A; finally, an eco category refers to the
categorization, in this case "A1.1", as described on wikipedia and of which
there are 13. Please read this commit message with this in mind.

I now made it possible to first do a vanilla search on all the categories.
When one then selects a certain category, it filters on the eco codes of
that category.

Then one can do a vanilla search on the classes, giving results in terms of
the alphabetic portions only.  Then one can select one of the classes, and
see the breakdown of that class in terms of its categories.

The eco codes portion stays basically the same.

Change-Id: I05440757767934e7902e630a956168a90ba4b5a0