Add the option to filter a given year
The last filter requirment is for a given year to be analyzed by minimum
elo value and by opening.  This option is now included.  It needs to be
tested a bit more thoroughly for actualling doing what I think it does.

The necessary query work is incorporated into query.php, hence,
eloc_query.php is deleted.

Change-Id: I5654a4c01c664cc9bf7c67a17d0009445cafcfb6