Make improvements to file uploading
Currently, if you upload a file, it does not get selected by default.  Now,
the adding of the file to the file selector is done with jquery, and only
gets done if file upload was successful.
Note, the behaviour for failing to upload file was not tested properly.

Finally, the responsive behavour of the graph was improved.

Change-Id: I0494e1d32a34fb396d2571107564a6c9325d36dd