Upgrade php/pgn_parser.php: whitespace and comments
Until now, we assumed the first seven tags, as well as the optional tags
wouldn't have whitespaces in between.  This updated now allows whitespace
to be present in any amount between lines and also validate the correctness
of tags before scanning them.

Many syntax errors in the millionbase file are unmasked now:
There are many (MANY) games in which no moves are indicated, only the score
where the moves should have been.  Quite a number of games have a deleted
first move line and just abrubly start move annotation on the sixth move.

While this is not critical for our implementation, any "additional" work we
might still consider might be influenced by this.

Change-Id: I107a535b079890b78791ce78383b6a4a9e5cd737