Handle single or no data point queries
When the user filters the data there is always the chance that they
specify the filters in such a way that the database only has 1 entry (or
none) that satisfies them. If there is only one entry, the visualisation
needs to be adapted (this happends in visual_response.js), if there is
no data returned, the server returns an error (this is in query.php).

Change-Id: I3742fa761d0820623357ccb790e995cf82f09368