Visualize gameplay with chessboard.js
This code is from the examples provided by the chesboardjs website.  It
uses the chess.js engine to check for valid moves.  Although I  don't know
the api's very well yet, I think we might be able to use this to have the
user enter moves through interactive game play, and have the computer make
suggestions based on our database.

Change-Id: I0142fbb476a20571424992edcbea77cd829af43e