Split php/query.php into four scripts
The query.php is handling up to 6 different kind of queries, with another
kind about to be added. This makes it horribly complicated to follow and
has scattered if statements all over the place.

Four new files are created, one for each kind of query required.
php/query.php itself is kept for now, but will probably soon be removed.
To avoid too much duplication and to aid in testing, much of the structure
is modularized into functions in php/utils.php.

Change-Id: I7c38275886ae9ab183ec8f6df01f84fa488b46f7