Add moves to database
This is quite a bit and confusing change, please go through it
carefully as I probably have made mistakes. This change adds the moves
data for games to the database as well. It implements the "star" method
as previously discussed.

A large problem with this update is that the creation of the database
now takes 10 times longer. This is actually not as bad as one would
expect (since there are about 40 moves to be inserted into the database
for each game. But it is still not great and we should investigate how
we can choose the batchsize for the moves to optimize for time.

My quick and messy choice was to use a moves batch size 2 times bigger
than the batch size for the tags data. I did this because the moves
table has about half as many columns as the tags table, thus one insert
into the tags table would take roughly the same time as 2 insters into
the moves table (I thiink?...).

I have also update the make files to install phpunit if it is not
present.

The exact number of move entries into the database is 101015624,
and the exact number of tag entries into the database is 2574755.

I have also found out what is causing the "non-valid" moves bug. It
seems that many moves (particularly between in 2014 and 2015) have no
moves recorded, just the result (eg: 0-1) in the place where the moves
should be. This is obviously not added to the moves table.

Change-Id: I7c73994ffbf2d26a06eac6a00f74e83118316950