Add details on optimal batch size to report.
I described the measuring of optimal batch size and the results found.  I
also tried to explain why an optimal batch size occurs.

Then, I gave a small breakdown on the effects an index table had on our
queries.

I removed metadata: this does not belong in a git repo.

When generating a pdf with pdflatex, it generates an aweful lot of metadata
in .log, .out and .aux files.  These should never be in the repository.  I
updated the .gitignore to not track them anymore and deleted them with this
commit.

Also, the actual pdf should not be in the repo, this can be generated
anywhere if you have the tex file.

Change-Id: If1503816e55eb316a6c1e93b73a5a276fae8db00